Adding query refactoring
